refactor: BookmarkList.tsx からコンポーネント・D&Dロジック・型定義を分離
786行の単一ファイルを7ファイルに分割し可読性・保守性を向上:
- types.ts: 共通型定義・定数
- DragHandleIcon.tsx, BookmarkItemContent.tsx, SortableBookmarkItem.tsx: UIコンポーネント
- CategoryDropZone.tsx, CategoryGroup.tsx: カテゴリグループコンポーネント
- useDragAndDrop.ts: D&Dロジックのカスタムフック
