+++ src/app/(dashboard)/bookmarks/types.ts
@@ -1,0 +1,20 @@
+export const UNCATEGORIZED_KEY = "__uncategorized__";
+
+export type Bookmark = {
+  id: string;
+  url: string;
+  title: string;
+  memo: string | null;
+  ogImage: string | null;
+  sortOrder: number;
+  tag: { id: string; name: string } | null;
+  tagId: string | null;
+};
+
+export type TagItem = { id: string; name: string };
+export type TagWithCount = { id: string; name: string; bookmarkCount: number };
+
+export type PendingDelete = {
+  bookmarks: Bookmark[];
+  timerId: ReturnType<typeof setTimeout>;
+};
